iota_sdk_move_types::iota_system::staking_pool

Struct StakingPoolV1

pub struct StakingPoolV1 {
    pub id: UID,
    pub activation_epoch: Option<u64>,
    pub deactivation_epoch: Option<u64>,
    pub iota_balance: u64,
    pub rewards_pool: Balance<IOTA>,
    pub pool_token_balance: u64,
    pub exchange_rates: Table<u64, PoolTokenExchangeRate>,
    pub pending_stake: u64,
    pub pending_total_iota_withdraw: u64,
    pub pending_pool_token_withdraw: u64,
    pub extra_fields: Bag,
}

Rust version of the Move iota_system::staking_pool::StakingPoolV1 type.

A staking pool embedded in each validator struct in the system state object.

Fields
id: UIDactivation_epoch: Option<u64>

The epoch at which this pool became active.

None while pre-active; Some(epoch) once active or inactive.

deactivation_epoch: Option<u64>

The epoch at which this pool ceased to be active.

None while pre-active or active; Some(epoch) once inactive.

iota_balance: u64

The total number of IOTA tokens in this pool — including the rewards_pool and the principal of every StakedIota — updated at epoch boundaries.

rewards_pool: Balance<IOTA>

Epoch stake rewards are added here at the end of each epoch.

pool_token_balance: u64

Total number of pool tokens issued by the pool.

exchange_rates: Table<u64, PoolTokenExchangeRate>

Exchange-rate history keyed by epoch number, starting from the activation_epoch and capturing the rate at the start of each epoch (after the previous epoch’s rewards are deposited).

pending_stake: u64

Pending stake amount for this epoch, emptied at epoch boundaries.

pending_total_iota_withdraw: u64

Pending stake withdrawn during the current epoch, including both principal and rewards; emptied at epoch boundaries.

pending_pool_token_withdraw: u64

Pending pool tokens withdrawn during the current epoch, emptied at epoch boundaries.

extra_fields: Bag

Any extra fields not defined statically.

Implementations

impl StakingPoolV1

pub const fn new( id: UID, activation_epoch: Option<u64>, deactivation_epoch: Option<u64>, iota_balance: u64, rewards_pool: Balance<IOTA>, pool_token_balance: u64, exchange_rates: Table<u64, PoolTokenExchangeRate>, pending_stake: u64, pending_total_iota_withdraw: u64, pending_pool_token_withdraw: u64, extra_fields: Bag, ) -> Self