iota_sdk_move_types

Trait MoveObject

pub trait MoveObject: Sized + for<'a> TryFrom<&'a Object, Error = FromObjectError> {
    // Required method
    fn struct_tag() -> StructTag;
}

Available on crate feature serde only.

A Rust mirror of a Move object that can be decoded from an Object.

struct_tag is the type of the objects it decodes, e.g. 0x2::coin::Coin<0x2::iota::IOTA> for Coin<IOTA>. Every MoveObject is also a MoveType with that same tag.

Required Methods

fn struct_tag() -> StructTag

The Move struct tag of the objects this type mirrors.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".